+++ get-latest.ps1
@@ -29,24 +29,34 @@ if ($XVM_VERSION) {
     }
     else {
         $XVM_RELEASE = "xvm-$XVM_VERSION.zip"
-        $XVM_URL = "https://dl1.modxvm.com/bin/$XVM_RELEASE"
+        $XVM_URL = "http://dl1.modxvm.com/get/bin/$XVM_RELEASE"
     }
-    if (test-path "latest") {
+    if (Test-Path "latest") {
         Remove-Item -recurse -force latest
     }
-    Write-Output "Downloading '$XVM_RELEASE'..."
-    Invoke-WebRequest -uri $XVM_URL -out "$XVM_RELEASE"
-    Write-Output "Extracting..."
-    Expand-Archive "$XVM_RELEASE" -DestinationPath latest
-    Write-Output "Clearing old PYC-files..."
-    $scriptDir = $(Split-Path $MyInvocation.MyCommand.Path)
     if ($verbose) {
-        Write-Output "[scriptDir = $scriptDir]"
+        Write-Output "Downloading '$XVM_URL'..."
     }
-    Get-ChildItem -Path "$scriptDir/../.." -Filter *.pyc -Recurse | ForEach-Object ($_) {Remove-Item $_.FullName}
-    Remove-Item "$scriptDir/../mods" -Recurse -Force
-§    Copy-Item -Recurse "$scriptDir/latest/res_mods/mods" "$scriptDir/../mods"
-    Write-Output "Launch Beyond Compare..."
-    bcomp . ./latest/res_mods/configs
-    Write-Output "Done."
+    else {
+        Write-Output "Downloading '$XVM_RELEASE'..."
+    }
+    Invoke-WebRequest -uri $XVM_URL -out "$XVM_RELEASE"
+    if (Test-Path $XVM_RELEASE) {
+        Write-Output "Extracting..."
+        Expand-Archive "$XVM_RELEASE" -DestinationPath latest
+        Write-Output "Clearing old PYC-files..."
+        $scriptDir = $(Split-Path $MyInvocation.MyCommand.Path)
+        if ($verbose) {
+            Write-Output "[scriptDir = $scriptDir]"
+        }
+        Get-ChildItem -Path "$scriptDir/../.." -Filter *.pyc -Recurse | ForEach-Object ($_) {Remove-Item $_.FullName}
+        Remove-Item "$scriptDir/../mods" -Recurse -Force
+        Copy-Item -Recurse "$scriptDir/latest/res_mods/mods" "$scriptDir/../mods"
+        Write-Output "Launch Beyond Compare..."
+        bcomp . ./latest/res_mods/configs
+        Write-Output "Done."
+    }
+    else {
+        Write-Output "Download failed."
+    }
 }
